move magentic and openai assistant agents
