--- python/packages/autogen-ext/src/autogen_ext/agents/__init__.py
@@ -1,4 +1,0 @@
-from ._magentic_one_coder_agent import MagenticOneCoderAgent
-from ._openai_assistant_agent import OpenAIAssistantAgent
-
-__all__ = ["OpenAIAssistantAgent", "MagenticOneCoderAgent"]
