+++ python/packages/autogen-ext/src/autogen_ext/agents/magentic/__init__.py
@@ -1,0 +1,3 @@
+from ._magentic_one_coder_agent import MagenticOneCoderAgent
+
+__all__ = ["MagenticOneCoderAgent"]
