+++ python/packages/autogen-ext/src/autogen_ext/agents/magentic/_magentic_one_coder_agent.py
