+++ python/packages/autogen-ext/src/autogen_ext/agents/openai/__init__.py
@@ -1,0 +1,3 @@
+from ._openai_assistant_agent import OpenAIAssistantAgent
+
+__all__ = ["OpenAIAssistantAgent"]
