+++ python/packages/autogen-ext/tests/test_openai_assistant_agent.py
@@ -6,7 +6,7 @@ import pytest
 from autogen_agentchat.messages import TextMessage
 from autogen_core import CancellationToken
 from autogen_core.components.tools._base import BaseTool, Tool
-from autogen_ext.agents import OpenAIAssistantAgent
+from autogen_ext.agents.openai import OpenAIAssistantAgent
 from azure.identity import DefaultAzureCredential, get_bearer_token_provider
 from openai import AsyncAzureOpenAI
 from pydantic import BaseModel
